fix(fsmv2/examplepanic): rename package example_panic → examplepanic
Go convention: package name must be lowercase without underscores and
should match the directory name. All three top-level files in the
examplepanic/ directory used the underscored form; align them.
