+++ umh-core/pkg/fsmv2/workers/example/examplepanic/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package example_panic
+package examplepanic
 
 import "github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsmv2/config"
